# Conflicts:
#	src/it/enne/curie/client/Client.java
#	src/it/enne/curie/common/CustomExtension.java
#	src/it/enne/curie/common/LogWriter.java
#	src/it/enne/curie/server/Server.java

+++ src/it/enne/curie/common/LogWriter.java
@@ -26,32 +26,42 @@ public class LogWriter {
     }
 
     public void write(String line) {
-        log.setWritable(true,true);
-        if (rowCont>=20) { // eliminazione prima riga file
-            File temp = new File(log+"tmp");
-            try (
-                    PrintWriter writer = new PrintWriter(new BufferedWriter(new OutputStreamWriter(new FileOutputStream(temp))));
-                    BufferedReader br = new BufferedReader(new FileReader(log))
-            ) {
-                String let;
-                br.readLine();
-                while ((let = br.readLine()) != null) {
-                    writer.write(let+"\n");
+        boolean isWritable;
+
+        isWritable = log.setWritable(true,true);
+        if (isWritable) {
+            if (rowCont >= 20) { // eliminazione prima riga file
+                File temp = new File(log + "tmp");
+                try (
+                        PrintWriter writer = new PrintWriter(new BufferedWriter(new OutputStreamWriter(new FileOutputStream(temp))));
+                        BufferedReader br = new BufferedReader(new FileReader(log))
+                ) {
+                    String let;
+                    br.readLine();
+                    while ((let = br.readLine()) != null) {
+                        writer.write(let + "\n");
+                    }
+                    rowCont--;
+                } catch (Exception e) {
+                    System.err.println("Errore eliminazione prima riga file log");
                 }
-                rowCont--;
+                if (!log.delete() || !temp.renameTo(log)) {
+                    System.err.println("errore switch  file temporaneo file log");
+                }
+            }
+            // aggiunta riga in fondo al file di log
+            try (PrintWriter writer = new PrintWriter(new BufferedWriter(new OutputStreamWriter(new FileOutputStream(log, true))))) {
+                writer.append(line).append("\n");
+                rowCont++;
             } catch (Exception e) {
-                System.err.println("Errore eliminazione prima riga file log");
+                System.err.println("Errore scrittura nel log: " + line);
             }
-            if (!log.delete() || !temp.renameTo(log)) {
-                System.err.println("errore switch  file temporaneo file log");
+            isWritable = log.setWritable(false, false);
+            if (!isWritable) {
+                System.err.println("log modificabile");
             }
-        }
-        // aggiunta riga in fondo al file di log
-        try (PrintWriter writer = new PrintWriter(new BufferedWriter(new OutputStreamWriter(new FileOutputStream(log, true))))) {
-            writer.append(line).append("\n");
-            rowCont++;
-        } catch (Exception e) {
-            System.err.println("Errore scrittura nel log: " + line);
+        } else {
+            System.err.println("errore scrittura log");
         }
         log.setWritable(false,false);
     }
